merges visiblecartlist and cart, recreates functions as selectors
